requirements.txt
# CoincallTrader Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
requests>=2.31.0          # HTTP client for API calls
python-dotenv>=1.0.0      # Environment variable management
APScheduler>=3.10.0       # Job scheduling

# Optional - for WebSocket support (future)
websockets>=12.0          # WebSocket client

# Development dependencies (optional)
# pytest>=7.0.0           # Testing framework
# pytest-asyncio>=0.21.0  # Async test support
